+++ src/helpers/runCommand.ts
@@ -1,17 +1,43 @@
+import path from 'path'
 import { spawn } from 'child_process'
-import type { ChildProcess } from 'child_process'
-import type { CLIFlags } from '@eswatch/types'
+import { getCLIOptions, parseOptions } from '@eswatch/helpers'
 
-type RunOutput = (options: CLIFlags) => ChildProcess
+const options = getCLIOptions()
 
-const runCommand: RunOutput = options => {
-  const commandToRun =
-    typeof options.run === 'string'
-      ? options.run
-      : `node --experimental-specifier-resolution=node ${options.outdir}`
+const isIndexFile = (file: string) => {
+  const filename = path.parse(file).name
+  return filename === 'index'
+}
+
+const getScriptWithExtension = (file: string) => {
+  const filename = path.parse(file).name
+  const extension = options.outext || '.js'
+  const filenameWithExtension = filename + extension
+  const filePath = path.resolve(options.outdir, filenameWithExtension)
+  return filePath
+}
+
+const getScriptToBeRan = (entryPoints: string[]): any => {
+  const indexFile = entryPoints.find(isIndexFile)
+  const scriptToBeRan = !indexFile ? entryPoints[0] : indexFile
+  return getScriptWithExtension(scriptToBeRan)
+}
+
+const getCommandToBeRan = (entryPoints: string[]) => {
+  if (typeof options.run === 'string') {
+    return options.run
+  } else {
+    const scriptToRun = getScriptToBeRan(entryPoints)
+    return 'node ' + scriptToRun
+  }
+}
+
+const runCommand = async () => {
+  const { entryPoints } = await parseOptions()
+  const commandToRun = getCommandToBeRan(entryPoints)
   const [commandName] = commandToRun.split(' ')
   const commandParameters = commandToRun.split(' ').slice(1)
-  // Consider replacing this `spawn` with a `fork`
+  // Consider replacing this `spawn` with a `fork` when --run is not string
   const childProcess = spawn(commandName, commandParameters, { stdio: 'inherit' })
   return childProcess
 }