+++ src/initEsbuild.ts
@@ -5,16 +5,16 @@ import { getCLIOptions, getEsbuildConfig } from '@eswatch/helpers'
 
 const initEsbuild = async () => {
   const options = getCLIOptions()
-  // Change return value of getEsbuildConfig to local types
   const esbuildConfig = await getEsbuildConfig()
   !options.keepfiles && rimraf.sync(esbuildConfig.outdir)
   const esbuildService = options.watch && (await esbuild.startService())
-  const buildOrWatch = (esbuildService ? esbuildService.build : esbuild.build).bind(
-    this,
-    esbuildConfig
-  )
-  const buildSteps: Promise<any>[] = [buildOrWatch(), postBuild()]
-  return async () => await Promise.all(buildSteps)
+  const buildOrWatch = (esbuildService ? esbuildService : esbuild).build.bind(this, esbuildConfig)
+  const buildSteps = [buildOrWatch, postBuild]
+  return async () => {
+    for (const buildStep of buildSteps) {
+      await buildStep()
+    }
+  }
 }
 
 export { initEsbuild }